DAFNI-wrappers/base-image/dockerFiles/requirements.txt
awscli==1.17.10
boto3== 1.11.10
fiona==1.8.9.post2
ipywidgets==7.5.1
jupyter==1.0.0
matplotlib==3.1.3
networkx==2.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
psycopg2==2.8.4
pyproj==2.4.2.post1
pyscaffold==3.2.2
pysftp==0.2.9
pytest==5.0.1
pyyaml==5.2
requests==2.22.0
rtree==0.9.3
scikit-learn==0.22.1
scipy==1.4.1
shapely==1.7.0
smif==1.3.1
urllib3==1.23
web.py==0.40
